Implement pyproject.toml for PyPI deployment and restructure project folders. Removed unnecessary dependencies and package-related files accordingly.
