--- requirements.txt
@@ -1,66 +1,0 @@
-annotated-types==0.7.0
-anyio==4.4.0
-appnope==0.1.4
-asttokens==2.4.1
-azure-ai-vision-imageanalysis==1.0.0b2
-azure-cognitiveservices-vision-computervision==0.9.0
-azure-common==1.1.28
-azure-core==1.30.2
-certifi==2024.6.2
-charset-normalizer==3.3.2
-comm==0.2.2
-contourpy==1.2.1
-cycler==0.12.1
-debugpy==1.8.1
-decorator==5.1.1
-distro==1.9.0
-executing==2.0.1
-fonttools==4.53.0
-h11==0.14.0
-httpcore==1.0.5
-httpx==0.27.0
-idna==3.7
-ipykernel==6.29.4
-ipython==8.25.0
-isodate==0.6.1
-jedi==0.19.1
-jupyter_client==8.6.2
-jupyter_core==5.7.2
-kiwisolver==1.4.5
-matplotlib==3.9.0
-matplotlib-inline==0.1.7
-msrest==0.7.1
-nest-asyncio==1.6.0
-numpy==1.25.2
-oauthlib==3.2.2
-openai==1.33.0
-opencv-python==4.10.0.82
-packaging==24.1
-parso==0.8.4
-pexpect==4.9.0
-pillow==10.3.0
-platformdirs==4.2.2
-prompt_toolkit==3.0.47
-psutil==5.9.8
-ptyprocess==0.7.0
-pure-eval==0.2.2
-pydantic==2.7.3
-pydantic_core==2.18.4
-Pygments==2.18.0
-pyparsing==3.1.2
-pytest==8.3.2
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-pyyaml==6.0.2
-pyzmq==26.0.3
-requests==2.32.3
-requests-oauthlib==2.0.0
-six==1.16.0
-sniffio==1.3.1
-stack-data==0.6.3
-tornado==6.4.1
-tqdm==4.66.4
-traitlets==5.14.3
-typing_extensions==4.12.2
-urllib3==2.2.1
-wcwidth==0.2.13
