--- setup.cfg
@@ -1,32 +1,0 @@
-[metadata]
-name = co_op_translator
-version = 0.1.0
-description = An open-source project for translating markdown files and images using Azure and OpenAI services
-
-[options]
-packages = find:
-package_dir =
-    = src
-install_requires =
-    azure-ai-vision-imageanalysis>=1.0.0b2, <2.0.0
-    azure-cognitiveservices-vision-computervision>=0.9.0, <1.0.0
-    azure-core>=1.30.2, <2.0.0
-    openai>=1.33.0, <2.0.0
-    opencv-python>=4.10.0.82, <5.0.0
-    Pillow>=10.3.0, <11.0.0
-    numpy>=1.25.2, <2.0.0
-    python-dotenv>=1.0.1, <2.0.0
-    requests>=2.32.3, <3.0.0
-    semantic-kernel>=1.8.2, <2.0.0
-
-python_requires = >=3.10
-
-[options.packages.find]
-where = src
-
-[options.entry_points]
-console_scripts =
-    translate = main:main
-
-[egg_info]
-egg_base = .
