+++ src/co_op_translator/config/constants.py
