+++ src/co_op_translator/config/font_language_mappings.yml
